Renamed sanic occurences with swordHero or swhero
