+++ public/index.html
@@ -8,12 +8,12 @@
     <meta name="viewport" content="user-scalable=no, initial-scale=1.0, width=device-width">
     <link rel="stylesheet" type="text/css" href="css/index.css">
     <link rel="shortcut icon" href="favicon.ico" />
-    <title>sAnic</title>
+    <title>swordHero</title>
 </head>
 
 <body>
     <div>
-        <!-- <h1 id="title">sAnic game</h1> -->
+        <!-- <h1 id="title">swordHero</h1> -->
 
         <div id="main"></div>
