+++ src/mz/swhero/AssetLoader.js
