+++ src/mz/swhero/GlobalConfig.js
