+++ src/mz/swhero/StaticEnemy.js
